fix(scheduler): convert Status enum to a NodeState enum properly
