+++ scheduler/src/parser/mod.rs
@@ -1,3 +1,4 @@
 pub mod instance;
 pub mod port;
 pub mod resource;
+pub mod status;
