+++ scheduler/src/parser/port.rs
@@ -1,6 +1,4 @@
-use proto::scheduler::{InstanceStatus, NodeStatus, Port};
-
-use super::resource::ResourceParser;
+use proto::scheduler::Port;
 
 pub struct PortParser {}
 
@@ -41,42 +39,5 @@ impl PortParser {
                 destination: port.destination,
             })
             .collect()
-    }
-}
-
-pub struct StatusParser {}
-
-impl StatusParser {
-    /// It takes a `proto::agent::InstanceStatus` and returns an `InstanceStatus`
-    ///
-    /// Arguments:
-    ///
-    /// * `status`: proto::agent::InstanceStatus
-    ///
-    /// Returns:
-    ///
-    /// A new InstanceStatus struct
-    pub fn from_agent_instance_status(status: proto::agent::InstanceStatus) -> InstanceStatus {
-        InstanceStatus {
-            id: status.id,
-            status: status.status,
-            status_description: status.description,
-            resource: status.resource.map(ResourceParser::from_agent_resource),
-        }
-    }
-
-    /// It takes a `NodeStatus` struct and returns a `proto::controller::NodeStatus` struct
-    ///
-    /// Arguments:
-    ///
-    /// * `status`: NodeStatus - this is the status of the node.
-    pub fn to_controller_node_status(status: NodeStatus) -> proto::controller::NodeStatus {
-        proto::controller::NodeStatus {
-            id: status.id,
-            state: status.status,
-            status_description: status.status_description,
-            resource: status.resource.map(ResourceParser::to_controller_resource),
-            instances: vec![], // todo;
-        }
     }
 }
